+++ src/main/java/com/stratio/qa/specs/CommonG.java
@@ -73,6 +73,7 @@ import java.io.*;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.net.InetAddress;
+import java.net.URISyntaxException;
 import java.net.UnknownHostException;
 import java.nio.charset.StandardCharsets;
 import java.sql.Connection;
@@ -698,7 +699,9 @@ public class CommonG {
         String httpResponse = response.getResponseBody();
         List<Cookie> cookies = response.getCookies();
         HttpHeaders headers = response.getHeaders();
-        this.response = new HttpResponse(statusCode, httpResponse, cookies, headers);
+        InputStream bodyStream = response.getResponseBodyAsStream();
+
+        this.response = new HttpResponse(statusCode, httpResponse, cookies, headers, bodyStream);
     }
 
     /**
@@ -2843,7 +2846,7 @@ public class CommonG {
     /**
      * Check if host is reachable
      *
-     * @param host Host
+     * @param host    Host
      * @param timeOut timeout
      * @return True/False
      */